% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{anova.test}
\alias{anova.test}
\title{Analysis of the repeated measured data in Animal Science}
\usage{
\method{anova}{test}(
  data = data,
  num.aa = 1,
  n = 6,
  time.points = 6,
  subid = "Steer",
  group = "Group",
  time = "Time",
  resp.var = "citrulline",
  amino.names = "citrulline",
  interv.length = 7,
  num.method = 4,
  corStruct = "gen.ar1",
  hetero = TRUE,
  file = "rpaa_cit_data.csv"
)
}
\arguments{
\item{data}{dataset you want to analyze whose response variables are repeatedly measured}

\item{num.aa}{number of amino acids, the default is 1}

\item{n}{number of animals}

\item{time.points}{number of level of time factor}

\item{subid}{subject id in your data set}

\item{group}{group name in your data set}

\item{time}{time points at which the response variable are measured at}

\item{resp.var}{repeated measured response variable in your data set (repeated measured)}

\item{amino.names}{name of the amino acid (response variables)}

\item{interv.length}{interval length between two grids to draw a y-axis}

\item{num.method}{number of ways to fit linear mixed effects model
1) with standard var-cov structure,
2) with heterogeneous variance and generalized AR(1),
3) with baseline covariates,
4) based on normalized data using the percentage-changes from baselines}

\item{corStruct}{choice of correlation structures, e.g) "ar1", "gen.ar1", "compsymm", and "unstruct"}

\item{hetero}{TRUE for unequal variance for the within-subject errors.}

\item{file}{data file name: .cvs file.  The data file format has to be csv file and long format with four columns: subject_id, group_id, Time and Outcomes for the repeated measured data}
}
\value{
two tables and three plots
}
\description{
Analysis of the repeated measured data in Animal Science
}
\examples{

library(ARMIS)
data(pseudo_steer_data)

anova.test(data=data, num.aa=1, n=6, time.points=6, subid="Steer",
group="Group", time="Time", resp.var="citrulline", amino.names="citrulline",
interv.length=7, num.method=4, corStruct="gen.ar1", hetero=TRUE,
file="rpaa_cit_data.csv")
}
